Make generate_test_cases part of the ISA
