+++ src/isa/standard/puzzle.rs
@@ -1,10 +1,7 @@
 use crossterm::style::Stylize;
 use crossterm::{cursor, QueueableCommand};
 use rand::Rng;
-use rlua::prelude::*;
 use std::collections::VecDeque;
-use std::error::Error;
-use std::fs;
 use std::io::{self, Write};
 
 use super::vm::{VAL_CHAR_WIDTH, VAL_MAX, VAL_MIN};
@@ -157,45 +154,4 @@ impl Printable for PuzzleIO {
   fn print(&self) -> io::Result<()> {
     self.print_with_expected_outputs(self)
   }
-}
-
-///
-/// Load and run the Lua code to generate the puzzles
-///
-pub fn generate_test_cases(lua_file: &str, seed: u32, n: usize) -> Result<TestCaseSet, Box<dyn Error>> {
-  // Try to load the Lua code file into memory
-  let lua_code = fs::read_to_string(format!("levels/{}", lua_file))?;
-
-  // Generate and run the code within the Lua context
-  let test_cases = Lua::new().context::<_, LuaResult<TestCaseSet>>(|ctx| {
-    let globals = ctx.globals();
-
-    // Add the levels folder to the path
-    ctx
-      .load(&r#"package.path = "./levels/?.lua;" .. package.path"#)
-      .exec()?;
-
-    // Seed the random number generator
-    globals
-      .get::<_, LuaTable>("math")?
-      .get::<_, LuaFunction>("randomseed")?
-      .call::<_, ()>(seed)?;
-
-    // Load the script code
-    //  This should define a global function named "generateTestCase"
-    ctx.load(&lua_code).exec()?;
-
-    // Generate the test cases one-by-one
-    let generate_test_case: LuaFunction = globals.get("generateTestCase")?;
-    let test_cases = (0..n)
-      .map(|_| {
-        let (inputs, outputs): (Vec<i16>, Vec<i16>) = generate_test_case.call(())?;
-        Puzzle::new(inputs, outputs).map_err(LuaError::RuntimeError)
-      })
-      .collect::<Result<_, _>>()?;
-
-    Ok(test_cases)
-  })?;
-
-  Ok(test_cases)
 }
